add part number attribute to flElbow
The names and the labels of the parts are unique, that is why I cannot
store part number in them. The problem occure, when there are multipe items
with the same part number in document. To solve this problem I adde
an new attribute "PartNumber" to the Flamingo implimentation of the flElbow