improve CreatePartGui.moveFlamingoPartToSelection
Show messages when positioning of Flamingo parts fails.
Add a workaround to a PipeFromTable.create: add ports to
pipeFeatures.Pipe after creating. For some reasons it is empty
when I call moveFlamingoPartToSelection. The other fittings
have their Ports updated.

There are still problems with position of the fittings:
couplings and Bushing are rotated.